Subscribe on private pools inside useWebsocketSubscription
