+++ src/pages/LendPage/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useLoansPage';
-export * from './usePrivateLendingPools';
